Rename --refresh-interval to --duration, refactor localserver code
- Renamed --refresh-interval to -d/--duration using existing DurationParam
  validation (15 minutes to 12 hours)
- Extracted localserver code into s3_credentials/localserver.py module:
  - CredentialCache class for thread-safe credential caching
  - make_credential_handler for HTTP request handling
  - run_server function to start the server
- Moved localserver tests to tests/test_localserver.py
- Added documentation in docs/localserver.md
- Updated docs index to include localserver
