+++ tests/test_s3_credentials.py
@@ -1269,416 +1269,3 @@ def test_delete_objects_dry_run(moto_s3_populated, arg):
         assert "directory/two.txt" in result.output
         after_keys = get_keys()
         assert before_keys == after_keys
-
-
-# Tests for localserver command
-
-
-@pytest.mark.parametrize(
-    "value,expected",
-    (
-        ("30s", 30),
-        ("5m", 300),
-        ("1h", 3600),
-        ("60", 60),  # No suffix means seconds
-        ("120s", 120),
-        ("10m", 600),
-        ("2h", 7200),
-    ),
-)
-def test_refresh_interval_param_valid(value, expected):
-    from s3_credentials.cli import RefreshIntervalParam
-
-    param = RefreshIntervalParam()
-    assert param.convert(value, None, None) == expected
-
-
-@pytest.mark.parametrize(
-    "value,expected_error",
-    (
-        ("invalid", "Refresh interval must be of form 30s or 5m or 1h"),
-        ("abc123", "Refresh interval must be of form 30s or 5m or 1h"),
-        ("-5m", "Refresh interval must be of form 30s or 5m or 1h"),
-        ("", "Refresh interval must be of form 30s or 5m or 1h"),
-    ),
-)
-def test_refresh_interval_param_invalid(value, expected_error):
-    from s3_credentials.cli import RefreshIntervalParam
-    import click
-
-    param = RefreshIntervalParam()
-    with pytest.raises(click.exceptions.BadParameter) as exc_info:
-        param.convert(value, None, None)
-    assert expected_error in str(exc_info.value)
-
-
-def test_localserver_missing_refresh_interval():
-    runner = CliRunner()
-    result = runner.invoke(cli, ["localserver", "my-bucket"])
-    assert result.exit_code == 2
-    assert "Missing option '--refresh-interval'" in result.output
-
-
-def test_localserver_invalid_refresh_interval():
-    runner = CliRunner()
-    result = runner.invoke(
-        cli, ["localserver", "my-bucket", "--refresh-interval", "invalid"]
-    )
-    assert result.exit_code == 2
-    assert "Refresh interval must be of form 30s or 5m or 1h" in result.output
-
-
-def test_localserver_read_only_write_only_conflict():
-    runner = CliRunner()
-    result = runner.invoke(
-        cli,
-        [
-            "localserver",
-            "my-bucket",
-            "--refresh-interval",
-            "5m",
-            "--read-only",
-            "--write-only",
-        ],
-    )
-    assert result.exit_code == 1
-    assert "Cannot use --read-only and --write-only at the same time" in result.output
-
-
-def test_localserver_bucket_not_exists(mocker):
-    boto3 = mocker.patch("boto3.client")
-    boto3.return_value = Mock()
-    boto3.return_value.head_bucket.side_effect = botocore.exceptions.ClientError(
-        error_response={}, operation_name=""
-    )
-
-    runner = CliRunner()
-    result = runner.invoke(
-        cli, ["localserver", "nonexistent-bucket", "--refresh-interval", "5m"]
-    )
-    assert result.exit_code == 1
-    assert "Bucket does not exist: nonexistent-bucket" in result.output
-
-
-def test_credential_cache_generates_credentials(mocker):
-    from s3_credentials.cli import CredentialCache
-    import datetime
-
-    mock_iam = Mock()
-    mock_sts = Mock()
-
-    mock_sts.get_caller_identity.return_value = {"Account": "123456"}
-    mock_iam.get_role.return_value = {"Role": {"Arn": "arn:aws:iam::123456:role/test"}}
-    mock_sts.assume_role.return_value = {
-        "Credentials": {
-            "AccessKeyId": "AKIAIOSFODNN7EXAMPLE",
-            "SecretAccessKey": "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY",
-            "SessionToken": "session-token",
-            "Expiration": datetime.datetime(2025, 12, 16, 12, 0, 0),
-        }
-    }
-
-    cache = CredentialCache(
-        iam=mock_iam,
-        sts=mock_sts,
-        bucket="test-bucket",
-        permission="read-only",
-        prefix="*",
-        refresh_interval=300,
-        extra_statements=[],
-    )
-
-    credentials = cache.get_credentials()
-
-    assert credentials["AccessKeyId"] == "AKIAIOSFODNN7EXAMPLE"
-    assert credentials["SecretAccessKey"] == "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
-    assert credentials["SessionToken"] == "session-token"
-
-    mock_sts.assume_role.assert_called_once()
-    call_kwargs = mock_sts.assume_role.call_args[1]
-    assert call_kwargs["RoleArn"] == "arn:aws:iam::123456:role/test"
-    assert call_kwargs["RoleSessionName"] == "s3.read-only.test-bucket"
-    assert call_kwargs["DurationSeconds"] == 900  # min 15 minutes
-
-
-def test_credential_cache_caches_credentials(mocker):
-    from s3_credentials.cli import CredentialCache
-    import datetime
-
-    mock_iam = Mock()
-    mock_sts = Mock()
-
-    mock_sts.get_caller_identity.return_value = {"Account": "123456"}
-    mock_iam.get_role.return_value = {"Role": {"Arn": "arn:aws:iam::123456:role/test"}}
-    mock_sts.assume_role.return_value = {
-        "Credentials": {
-            "AccessKeyId": "AKIAIOSFODNN7EXAMPLE",
-            "SecretAccessKey": "secret",
-            "SessionToken": "token",
-            "Expiration": datetime.datetime(2025, 12, 16, 12, 0, 0),
-        }
-    }
-
-    cache = CredentialCache(
-        iam=mock_iam,
-        sts=mock_sts,
-        bucket="test-bucket",
-        permission="read-write",
-        prefix="*",
-        refresh_interval=300,
-        extra_statements=[],
-    )
-
-    # Get credentials twice
-    creds1 = cache.get_credentials()
-    creds2 = cache.get_credentials()
-
-    # Should be the same object (cached)
-    assert creds1 is creds2
-    # Should only have called assume_role once
-    assert mock_sts.assume_role.call_count == 1
-
-
-def test_credential_cache_refreshes_after_interval(mocker):
-    from s3_credentials.cli import CredentialCache
-    import datetime
-    import time
-
-    mock_iam = Mock()
-    mock_sts = Mock()
-
-    mock_sts.get_caller_identity.return_value = {"Account": "123456"}
-    mock_iam.get_role.return_value = {"Role": {"Arn": "arn:aws:iam::123456:role/test"}}
-    mock_sts.assume_role.return_value = {
-        "Credentials": {
-            "AccessKeyId": "AKIAIOSFODNN7EXAMPLE",
-            "SecretAccessKey": "secret",
-            "SessionToken": "token",
-            "Expiration": datetime.datetime(2025, 12, 16, 12, 0, 0),
-        }
-    }
-
-    cache = CredentialCache(
-        iam=mock_iam,
-        sts=mock_sts,
-        bucket="test-bucket",
-        permission="read-write",
-        prefix="*",
-        refresh_interval=1,  # 1 second
-        extra_statements=[],
-    )
-
-    # Get credentials first time
-    cache.get_credentials()
-    assert mock_sts.assume_role.call_count == 1
-
-    # Wait for interval to expire
-    time.sleep(1.1)
-
-    # Get credentials again - should regenerate
-    cache.get_credentials()
-    assert mock_sts.assume_role.call_count == 2
-
-
-@pytest.mark.parametrize(
-    "permission,expected_permission",
-    (
-        ("read-write", "read-write"),
-        ("read-only", "read-only"),
-        ("write-only", "write-only"),
-    ),
-)
-def test_credential_cache_permission_in_session_name(mocker, permission, expected_permission):
-    from s3_credentials.cli import CredentialCache
-    import datetime
-
-    mock_iam = Mock()
-    mock_sts = Mock()
-
-    mock_sts.get_caller_identity.return_value = {"Account": "123456"}
-    mock_iam.get_role.return_value = {"Role": {"Arn": "arn:aws:iam::123456:role/test"}}
-    mock_sts.assume_role.return_value = {
-        "Credentials": {
-            "AccessKeyId": "AKIAIOSFODNN7EXAMPLE",
-            "SecretAccessKey": "secret",
-            "SessionToken": "token",
-            "Expiration": datetime.datetime(2025, 12, 16, 12, 0, 0),
-        }
-    }
-
-    cache = CredentialCache(
-        iam=mock_iam,
-        sts=mock_sts,
-        bucket="my-bucket",
-        permission=permission,
-        prefix="*",
-        refresh_interval=300,
-        extra_statements=[],
-    )
-
-    cache.get_credentials()
-
-    call_kwargs = mock_sts.assume_role.call_args[1]
-    assert call_kwargs["RoleSessionName"] == f"s3.{expected_permission}.my-bucket"
-
-
-def test_credential_cache_policy_generation(mocker):
-    from s3_credentials.cli import CredentialCache
-    import datetime
-
-    mock_iam = Mock()
-    mock_sts = Mock()
-
-    mock_sts.get_caller_identity.return_value = {"Account": "123456"}
-    mock_iam.get_role.return_value = {"Role": {"Arn": "arn:aws:iam::123456:role/test"}}
-    mock_sts.assume_role.return_value = {
-        "Credentials": {
-            "AccessKeyId": "AKIAIOSFODNN7EXAMPLE",
-            "SecretAccessKey": "secret",
-            "SessionToken": "token",
-            "Expiration": datetime.datetime(2025, 12, 16, 12, 0, 0),
-        }
-    }
-
-    cache = CredentialCache(
-        iam=mock_iam,
-        sts=mock_sts,
-        bucket="test-bucket",
-        permission="read-only",
-        prefix="*",
-        refresh_interval=300,
-        extra_statements=[],
-    )
-
-    cache.get_credentials()
-
-    call_kwargs = mock_sts.assume_role.call_args[1]
-    policy = json.loads(call_kwargs["Policy"])
-    assert policy["Version"] == "2012-10-17"
-    assert len(policy["Statement"]) == 2
-    # Should have ListBucket and GetObject statements
-    actions = []
-    for stmt in policy["Statement"]:
-        actions.extend(stmt["Action"])
-    assert "s3:ListBucket" in actions
-    assert "s3:GetObject" in actions
-
-
-def test_make_credential_handler_returns_credentials(mocker):
-    from s3_credentials.cli import make_credential_handler
-    import datetime
-    import io
-
-    mock_cache = Mock()
-    mock_cache.get_credentials.return_value = {
-        "AccessKeyId": "AKIAIOSFODNN7EXAMPLE",
-        "SecretAccessKey": "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY",
-        "SessionToken": "session-token",
-        "Expiration": datetime.datetime(2025, 12, 16, 12, 0, 0),
-    }
-
-    handler_class = make_credential_handler(mock_cache)
-
-    # Create a mock request
-    mock_request = Mock()
-    mock_request.makefile.return_value = io.BytesIO(b"GET / HTTP/1.1\r\n\r\n")
-
-    # Create handler instance with mocked internals
-    handler = handler_class.__new__(handler_class)
-    handler.path = "/"
-    handler.wfile = io.BytesIO()
-    handler.request_version = "HTTP/1.1"
-    handler.requestline = "GET / HTTP/1.1"
-
-    # Mock response methods
-    response_code = None
-    headers = {}
-
-    def mock_send_response(code):
-        nonlocal response_code
-        response_code = code
-
-    def mock_send_header(name, value):
-        headers[name] = value
-
-    def mock_end_headers():
-        pass
-
-    handler.send_response = mock_send_response
-    handler.send_header = mock_send_header
-    handler.end_headers = mock_end_headers
-
-    # Call do_GET
-    handler.do_GET()
-
-    # Verify response
-    assert response_code == 200
-    assert headers["Content-Type"] == "application/json"
-
-    # Parse the response body
-    response_body = handler.wfile.getvalue().decode()
-    response_json = json.loads(response_body)
-    assert response_json["AccessKeyId"] == "AKIAIOSFODNN7EXAMPLE"
-    assert response_json["SecretAccessKey"] == "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
-    assert response_json["SessionToken"] == "session-token"
-
-
-def test_make_credential_handler_404_on_wrong_path(mocker):
-    from s3_credentials.cli import make_credential_handler
-    import io
-
-    mock_cache = Mock()
-    handler_class = make_credential_handler(mock_cache)
-
-    handler = handler_class.__new__(handler_class)
-    handler.path = "/wrong-path"
-    handler.wfile = io.BytesIO()
-    handler.request_version = "HTTP/1.1"
-
-    response_code = None
-    headers = {}
-
-    def mock_send_response(code):
-        nonlocal response_code
-        response_code = code
-
-    handler.send_response = mock_send_response
-    handler.send_header = lambda name, value: headers.update({name: value})
-    handler.end_headers = lambda: None
-
-    handler.do_GET()
-
-    assert response_code == 404
-    response_body = handler.wfile.getvalue().decode()
-    assert "Not found" in response_body
-
-
-def test_make_credential_handler_500_on_error(mocker):
-    from s3_credentials.cli import make_credential_handler
-    import io
-
-    mock_cache = Mock()
-    mock_cache.get_credentials.side_effect = Exception("AWS Error")
-
-    handler_class = make_credential_handler(mock_cache)
-
-    handler = handler_class.__new__(handler_class)
-    handler.path = "/"
-    handler.wfile = io.BytesIO()
-    handler.request_version = "HTTP/1.1"
-
-    response_code = None
-
-    def mock_send_response(code):
-        nonlocal response_code
-        response_code = code
-
-    handler.send_response = mock_send_response
-    handler.send_header = lambda name, value: None
-    handler.end_headers = lambda: None
-
-    handler.do_GET()
-
-    assert response_code == 500
-    response_body = handler.wfile.getvalue().decode()
-    assert "AWS Error" in response_body
